+++ src/platform/packages/shared/kbn-unified-tabs/src/components/tabs_visual_glue_to_header/constants.ts
@@ -7,4 +7,9 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export const MAX_TAB_LABEL_LENGTH = 500;
+export enum zLevels {
+  belowHeaderShadowTabBackground = 0,
+  underHeaderShadow = 1,
+  aboveHeaderShadowTabBackground = 2,
+  tabMainContent = 3,
+}
